refactor(server): #125 — extract KV store handlers into routes/stores.rs
